% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_it_rain.R
\name{make_it_rain}
\alias{make_it_rain}
\title{Generate Matrix (movie) code rain in the RStudio viewer}
\usage{
make_it_rain()
}
\description{
If you want to make it rain Matrix code, make_it_rain()
}
\examples{

# it's pretty easy to make it rain in the RStudio viewer
make_it_rain()

}
